requirements.txt
scrapegraphai
pandas
python-dotenv